bbu_reference/backtest_reference/bbu_backtest-main/pyproject.toml
[project]
name = "bbu-backtest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "pybit>=5.11.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.42",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.8",
    "pytest-cov>=6.2.1",
]
